[IMP] mail: improve quote reply detection
We update quote detection for gmail and outlook:
- gmail has simple wrapper divs with explicit classes
- outlook has a mix of div ids and simple pattern-based quoting
(everything under "<hr><div id="divRplyFwdMsg"/> seems to be considered a quote)

Previously gmail just used blockquote, which still works but does not
capture "On xx:xx:xx X <X@gmail.com> wrote:" headers, which are caught
for outlook.

Previously outlook had a wrapper div around divRplyFwdMsg which would
set data-o-mail-quote-container on it, and propagate to children.
However it seems that outer div was either removed or is not always
present, a heuristic is thus needed.

task-4381505

X-original-commit: 1b098fa17f953b24374c10ceb9f1753bba7dff7c
Part-of: odoo/odoo#194853
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Renaud Thiry (reth) <reth@odoo.com>