[FIX] pos_self_order_stripe: fix Stripe payments
Currently in v18 Stripe payments are broken and don't work anymore.
This PR adjusts the code so that Stripe works again in self order mode.
It fixes the error "An error has occured" seen when sending a payment to
a Stripe terminal + it also fixes the response sent to PoS to confirm
Stripe payments.

opw-4283413
opw-4358017

closes odoo/odoo#194823

X-original-commit: 0851fd095b11e4995d640b5aaeb3b2d11f94ac3b
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>
Signed-off-by: Yaroslav Soroko (yaso) <yaso@odoo.com>